Delete orca files

+++ routine.py
@@ -12,7 +12,6 @@ def read_write_trj(filename,i):
     with open(filename,"r") as file:
         lines = file.readlines()
         fi = open("water"+str(i)+".xyz","w")
-        print(lines[-5:], "water"+str(i)+".xyz")
         fi.writelines(lines[-5:])
 
 for i in range(0,100):
@@ -25,8 +24,11 @@ for i in range(0,100):
         os.system('runorca_4_2 stretched_water'+str(i-1)+'.inp')
         os.system('rm stretched_water'+str(i-1)+'.inp')
         read_write_trj('stretched_water'+str(i-1)+'_trj.xyz',i)
-        os.system('rm stretched_water'+str(i-1)+'_trj.xyz')
+        if(i<99):
+            os.system('rm stretched_water'+str(i-1)+'_trj.xyz')
+            os.system('rm stretched_water'+str(i-1)+'.out')
+            os.system('rm stretched_water'+str(i-1)+'.gbw')
         stretch_molecule('water'+str(i)+'.xyz')
         os.system('rm water'+str(i)+'.xyz')
         xyz_to_orca('stretched_water'+str(i)+'.xyz')
-        os.system('rm stretched_water'+str(i)+'.xyz')
+        # os.system('rm stretched_water'+str(i)+'.xyz')
